Sign compare patch (#13)
Patched build failure against upstream pytorch where no-sign-compare is
enforced now.

+++ csrc/python_frontend/fusion_record.h
@@ -1199,7 +1199,7 @@ struct OutputRecord : RecordFunctor {
     if (!stride_order.empty()) {
       bool requires_permutation = false;
       for (const auto i : c10::irange(stride_order.size())) {
-        if (stride_order[i] != i) {
+        if (stride_order[i] != (int64_t)i) {
           requires_permutation = true;
           break;
         }
@@ -1268,7 +1268,7 @@ struct OutputRecord : RecordFunctor {
 
         if (!stride_order_.empty()) {
           std::vector<int64_t> reverse_perm(stride_order_.size());
-          size_t duplicate_check = 0;
+          int64_t duplicate_check = 0;
           for (const auto i : c10::irange(stride_order_.size())) {
             TORCH_CHECK(
                 stride_order_[i] >= 0 &&
